delete protection etc

+++ controllers/cards.js
@@ -4,12 +4,11 @@ const ERROR_CODE = 400;
 const NOTFOUND_CODE = 404;
 const DEFAULTERROR_CODE = 500;
 
-//Я сам пытался давать неправильный id и когда как получал ReferenceError или CastError, наверно просто упустил ее, про все читал в документации mongoose'a
 const errorHandling = (err, res) => {
   switch (err.name) {
     case 'CastError':
     case 'ValidationError':
-      res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные.' });
+      res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные.', err });
       break;
     case 'ReferenceError':
       res.status(NOTFOUND_CODE).send({ message: 'Карточка с указанным _id не найдена.' });
@@ -22,26 +21,34 @@ const errorHandling = (err, res) => {
 
 export const getCards = (req, res) => {
   Card.find({})
+    .populate(['owner', 'likes'])
     .then((cards) => res.send(cards))
     .catch((err) => errorHandling(err, res));
 };
 
 export const createCards = (req, res) => {
-  const { _id } = req.user;
+  // const { _id } = req.user;
   const { name, link } = req.body;
-  Card.create({ name, link, owner: _id })
+  Card.create({ name, link, owner: req.user })
     .then((card) => res.send(card))
     .catch((err) => errorHandling(err, res));
 };
 
 export const deleteCards = (req, res) => {
-  Card.findByIdAndRemove({ _id: req.params.cardId })
-    .then((data) => {
-      if (data) {
-        res.send({ message: 'Карточка удалена' });
-      } else throw new ReferenceError();
-    })
-    .catch((err) => errorHandling(err, res));
+  Card.findById({ _id: req.params.cardId })
+    .then((card) => {
+      if (req.user._id === card.owner._id.toString()) {
+        Card.findByIdAndRemove({ _id: req.params.cardId })
+          .then((data) => {
+            if (data) {
+              res.send({ message: 'Карточка удалена' });
+            } else throw new ReferenceError();
+          })
+          .catch((err) => errorHandling(err, res));
+      } else {
+        res.status(403).send({ message: 'Вы не владелец карточки' });
+      }
+    });
 };
 
 export const likeCard = (req, res) => {
@@ -50,6 +57,7 @@ export const likeCard = (req, res) => {
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
+    .populate(['owner', 'likes'])
     .then((card) => {
       if (card) {
         res.send(card);
@@ -64,6 +72,7 @@ export const dislikeCard = (req, res) => {
     { $pull: { likes: req.user._id } },
     { new: true },
   )
+    .populate(['owner', 'likes'])
     .then((card) => {
       if (card) {
         res.send(card);
